+++ app/src/main/java/com/sopt/geonppang/presentation/signup/SignUpEmailActivity.kt
@@ -1,11 +1,10 @@
-package com.sopt.geonppang.presentation.signup.view
+package com.sopt.geonppang.presentation.signup
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import com.sopt.geonppang.R
 import com.sopt.geonppang.databinding.ActivitySignupEmailBinding
-import com.sopt.geonppang.presentation.signup.viewmodel.SignUpViewModel
 import com.sopt.geonppang.util.binding.BindingActivity
 import com.sopt.geonppang.util.extension.hideKeyboard
 import dagger.hilt.android.AndroidEntryPoint
