+++ app/src/main/java/com/sopt/geonppang/presentation/signup/SignUpNicknameActivity.kt
@@ -1,18 +1,16 @@
-package com.sopt.geonppang.presentation.signup.view
+package com.sopt.geonppang.presentation.signup
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import com.sopt.geonppang.R
-import com.sopt.geonppang.databinding.ActivitySignupPasswordBinding
-import com.sopt.geonppang.presentation.signup.viewmodel.SignUpViewModel
+import com.sopt.geonppang.databinding.ActivitySignupNicknameBinding
 import com.sopt.geonppang.util.binding.BindingActivity
 import com.sopt.geonppang.util.extension.hideKeyboard
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class SignUpPasswordActivity :
-    BindingActivity<ActivitySignupPasswordBinding>(R.layout.activity_signup_password) {
+class SignUpNicknameActivity :
+    BindingActivity<ActivitySignupNicknameBinding>(R.layout.activity_signup_nickname) {
     private val viewModel: SignUpViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -22,17 +20,15 @@ class SignUpPasswordActivity :
 
         addListeners()
     }
+
     private fun addListeners() {
         binding.root.setOnClickListener {
             hideKeyboard(it)
         }
-        binding.btnNext.setOnClickListener {
-            moveToNickname()
-        }
-    }
 
-    private fun moveToNickname() {
-        startActivity(Intent(this, SignUpNicknameActivity::class.java))
-        finish()
+        binding.btnDoubleCheck.setOnClickListener {
+            val bottomSheetDialog = SignUpNicknameBottomSheetDialog()
+            bottomSheetDialog.show(supportFragmentManager, bottomSheetDialog.tag)
+        }
     }
 }
