+++ app/src/main/java/com/sopt/geonppang/presentation/signup/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package com.sopt.geonppang.presentation.signup.viewmodel
+package com.sopt.geonppang.presentation.signup
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
@@ -68,8 +68,7 @@ class SignUpViewModel @Inject constructor() : ViewModel() {
 
     companion object {
         const val EMAIL_PATTERN = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+\$"
-        const val NICKNAME_PATTERN = "[가-힣]*[A-Za-z[0-9]]{2,10}\$"
-        const val PASSWORD_PATTERN =
-            "^(?=.*[A-Za-z])(?=.*[0-9])(?=.*[\$@\$!%*#?&.])[A-Za-z[0-9]\$@\$!%*#?&.]{6,12}\$"
+        const val NICKNAME_PATTERN = "[가-힣]*[A-Za-z[0-9]]{1,10}\$"
+        const val PASSWORD_PATTERN = "^[A-Za-z0-9]{8,25}$"
     }
 }
